p3.1.advanced-deps/myfunction/requirements.txt
PyYaml>=5.1,<6.0
markdown2==2.3.5
natsort==5.4.1
